requirements-legacy.txt
# Core dependencies
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
tqdm>=4.65.0

# Configuration
python-dotenv>=1.0.0
pydantic>=2.0.0

# Data generation
mdatagen>=0.2.0
missmecha-py>=0.1.2

# Gradient Boosting optimized
xgboost>=2.0.0
catboost>=1.2.0
optuna>=3.4.0

# LLM (optional - for feature extraction with AI)
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-google-genai>=0.0.5
